# payment/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Payment Application API
  version: 1.0.0
paths:
  /api/payments:
    post:
      tags:
        - payment
      operationId: createPayment
      summary: Create a new payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201':
          description: Payment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringResponse'
        '400':
          description: Validation error
        '500':
          description: Internal server error

  /api/webhooks:
    post:
      tags:
        - webhook
      operationId: registerWebhook
      summary: Register a new webhook URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookRequest'
      responses:
        '201':
          description: Webhook registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringResponse'
        '400':
          description: Validation error
        '500':
          description: Internal server error
components:
  schemas:

    StringResponse:
      type: object
      properties:
        message:
          type: string
          example: "Operation completed successfully"

    PaymentRequest:
      type: object
      properties:
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        zip:
          type: string
          example: "12345"
        cardNumber:
          type: string
          example: "4111111111111111"
        amount:
          type: number
          example: 100.50
        clientId:
          type: string
          example: "client_123"

    WebhookRequest:
      type: object
      properties:
        url:
          type: string
          example: "http://localhost:8080/dummywebhook/receive"
        clientId:
          type: string
          example: "client_123"